Big changes to Player
Commented and cleaned up.
nextCard now takes an Entity as a parameter(the target), and returns a boolean value instead of a Card; true means the player hasn't ended there turn and the loop in Game will continue, and false means the user has ended their turn. nextCard now plays the card instead of Game, and shows the player the card's stats before using the Card or cancelling usage.